backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pytest
pytest-asyncio
httpx
python-dotenv

